Rename opg-pipeline-builder folder to 'src' in prep for uv
