+++ src/command_line.py
@@ -3,10 +3,10 @@ import os
 import sys
 from copy import deepcopy
 
-from opg_pipeline_builder import __version__ as v
-from opg_pipeline_builder.pipeline import Pipeline
-from opg_pipeline_builder.pipeline_builder import PipelineBuilder
-from opg_pipeline_builder.utils.constants import etl_steps
+from src import __version__ as v
+from src.pipeline import Pipeline
+from src.pipeline_builder import PipelineBuilder
+from src.utils.constants import etl_steps
 
 
 class bcolors:
