+++ tests/test_schema_reader.py
@@ -1,14 +1,16 @@
-from arrow_pd_parser import writer
 from tempfile import NamedTemporaryFile
-from mojap_metadata import Metadata
-from tests.helpers import mock_get_file, set_up_s3, dummy_bucket
+
 import pandas as pd
 import pytest
+from arrow_pd_parser import writer
+from mojap_metadata import Metadata
+
+from tests.helpers import dummy_bucket, mock_get_file, set_up_s3
 
 
 class TestSchemaReader:
-    import opg_pipeline_builder.utils.schema_reader as srutils
-    from opg_pipeline_builder.utils.constants import aws_region
+    import src.utils.schema_reader as srutils
+    from src.utils.constants import aws_region
 
     sr_utils = srutils
     aws_region = aws_region
